+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release 1.1 (UPCOMING RELEASE)
 
 * All money types automatically format on blur:  values will be updated, if necessary, to always end in a period followed by two digits
-* Tightened up the float types to not allow more than one leading zero, and to disallow negative zero
+* Relaxed the int and float types/subtypes, and added strict versions of each.  The relaxed versions accept leading/trailing zeros and negative zero; the strict versions do not.
 * Added a logging callback function as a configurable option
 * Added unit tests